+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * `ContinuousRange` and `DiscreteRange` methods now properly inherit and are fully
   mutable (@dpseidel).
+  
+* `log_breaks()` provides usable breaks even with very small ranges
+  (@billdenney, #168)
 
 # scales 1.0.0